% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_class}
\alias{add_class}
\title{Add class}
\usage{
add_class(x, class_to_add = "exp")
}
\arguments{
\item{x}{An object}

\item{class_to_add}{String of the class to add, defaults to "exp"}
}
\description{
This function adds a given class to an object, so that
different S3 methods can be called (e.g., `tidy.exp` to get OR for logistic regression models
(e.g., in `modelsummary::msummary`)
}
\keyword{internal}
