% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{add_package_snippets}
\alias{add_package_snippets}
\title{Export package-specific R snippets}
\source{
https://stackoverflow.com/a/62223103/10581449
}
\usage{
add_package_snippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snipped definitions were added
}
\description{
\code{add_package_snippets} copies all (missing) snippet definitions
  in 'inst/rstudio/r.snippets' and 'rmd.snippets' (if not empty) to the RStudio user snippet location.
}
\examples{
\dontrun{
add_package_snippets()
}
}
\keyword{internal}
