% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{apa_cor_table}
\alias{apa_cor_table}
\title{Create a correlation table with summary statistics in APA style}
\source{
Based on the apaTables \code{apa.cor.table()} function, but adapted to
accept weighted correlation matrices and work with the `gt` package instead. Code for calculation
of confidence intervals adapted from https://medium.com/@shandou/how-to-compute-confidence-interval-for-pearsons-r-a-brief-guide-951445b9cb2d`
}
\usage{
apa_cor_table(
  cor_matrix,
  ci = c("given", "z_transform", "simple_SE"),
  n = NULL,
  add_distributions = FALSE,
  data = NULL,
  filename = NULL,
  notes = list(NULL),
  stars = NULL,
  add_title = FALSE,
  extras = NULL,
  apa_style = TRUE,
  ...
)
}
\arguments{
\item{cor_matrix}{A correlation matrix, for example returned from
\code{cor_matrix()}, \code{svy_cor_matrix()}, or \code{wtd_cor_matrix_mi()}}

\item{ci}{Method to create CI - default is to use any given in the cor_matrix,
and otherwise to compute them using z-transformations. The simple SE method should not be used, but is provided for compatibility.}

\item{n}{Number of observations to calculate confidence intervals - only needed
if cor_matrix does not contain degrees of freedom (df) and confidence intervals are to
be calulcated using z-transformations}

\item{add_distributions}{Add graphs showing variable distributions?}

\item{data}{Original data, only needed if \code{add_distribution = TRUE}}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{notes}{List of additional notes to show under the table.}

\item{stars}{A character vector to change the significance symbols (see details in `sigstars`)}

\item{add_title}{Should title be added to table? Set to TRUE for default title or to character for custom title}

\item{extras}{Tibble of additional columns to be added after the descriptives column - needs to be sorted in the same order as the `desc` element in the cor_matrix}

\item{apa_style}{Logical, should APA-style formatting be applied}

\item{...}{
  Arguments passed on to \code{\link[=plot_distributions]{plot_distributions}}
  \describe{
    \item{\code{plot_type}}{Type of plot that should be produced - `histogram` or `density` plot. If `auto`,
histograms are produced for variables that take fewer than 10 unique values, density plots for others. If a number is provided,
that number is used as the maximum number of unique values for which a histogram is used.}
    \item{\code{hist_align_y}}{Should histograms use the same y-axis, so that bin heights are comparable? Defaults to FALSE}
    \item{\code{plot_theme}}{Additional theme_ commands to be added to each plot}
  }}
}
\value{
A table that can be printed in the RStudio console to be shown in the
viewer. Unless it is to be post-processed with further `gt` functions, it should
usually be saved by passing a filename argument.
}
\description{
This function, creates (and optionally saves) a correlation table with
summary statistics. It accepts correlation matrices from various functions
in this package as its first argument
}
