% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.fetch.GENE.grch37.mart}
\alias{biomart.fetch.GENE.grch37.mart}
\title{Uses biomaRt function to fetch the Ensembl Gene DB}
\usage{
biomart.fetch.GENE.grch37.mart()
}
\value{
BioMart Gene database (grch37 built, homo sapients)
}
\description{
Uses biomaRt function to fetch the Ensembl Gene DB (grch37 built, homo sapients)
}
\examples{
mybiomart <- biomart.fetch.GENE.grch37.mart()
}
