% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.fetch.SNP.grch37.mart}
\alias{biomart.fetch.SNP.grch37.mart}
\title{Uses biomaRt function to fetch the Ensembl SNP DB}
\usage{
biomart.fetch.SNP.grch37.mart()
}
\value{
BioMart SNP database (grch37 built, homo sapients)
}
\description{
Uses biomaRt function to fetch the Ensembl SNP DB (grch37 built, homo sapients)
}
\examples{
mybiomart <- biomart.fetch.SNP.grch37.mart()
}
\author{
Alessia Visconti
}
