% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.gene.in.window}
\alias{biomart.gene.in.window}
\title{Fetches the HUGO gene symbols in the spefified window}
\usage{
biomart.gene.in.window(chr, start, end, mart, with.position = FALSE)
}
\arguments{
\item{chr}{chromosome number (1, 2, ...)}

\item{start}{fist position of the window (included)}

\item{end}{last position of the window (included)}

\item{mart}{BioMart database}

\item{with.position}{whether included chromosome name and start and end position (default: false)}
}
\value{
HUGO gene symbols, or HUGO gene symbols plus gene boudaries
}
\description{
Uses biomaRt function to fetch the HUGO gene symbols for those genes included in the 
specified window. It uses the given BioMart database
}
\examples{
mybiomart <- biomart.fetch.GENE.grch37.mart()
biomart.gene.in.window(1, 1, 20000, mybiomart)
}
\author{
Alessia Visconti
}
