% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.gene.position}
\alias{biomart.gene.position}
\title{Fetches the chromosomal coordinates of the specified gene}
\usage{
biomart.gene.position(gene, mart)
}
\arguments{
\item{gene}{the variant mame (HUGO gene symbols)}

\item{mart}{BioMart database}
}
\value{
genomic coordinates
}
\description{
Uses biomaRt function to fetch the chromosomal coordinates for the given 
HUGO gene symbols. It uses the given BioMart database.
}
\examples{
mybiomart <- biomart.fetch.GENE.grch37.mart()
biomart.gene.position("DDX11L1", mybiomart)
}
\author{
Alessia Visconti
}
