% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.SNPid.in.window}
\alias{biomart.SNPid.in.window}
\title{Fetches the reference SNP IDs in the spefified window}
\usage{
biomart.SNPid.in.window(chr, start, end, mart)
}
\arguments{
\item{chr}{chromosome number (1, 2, ...)}

\item{start}{fist position of the window (included)}

\item{end}{last position of the window (included)}

\item{mart}{BioMart database}
}
\value{
reference SNP IDs
}
\description{
Uses biomaRt function to fetch the reference SNP IDs for those SNPs included in the 
specified window. It uses the given BioMart database
}
\examples{
mybiomart <- biomart.fetch.SNP.grch37.mart()
biomart.SNPid.in.window(1, 1, 10100, mybiomart)
}
\author{
Alessia Visconti
}
