% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.floor_ceiling}
\alias{.floor_ceiling}
\title{Helper function to round up and down in turn}
\usage{
.floor_ceiling(x, i)
}
\arguments{
\item{x}{Numeric, to be rounded}

\item{i}{Iterator. floor() will be used on x for odd i, ceiling() for even i}
}
\description{
Iterates between floor() and ceiling()
}
\keyword{internal}
