% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.is.numeric_col}
\alias{.is.numeric_col}
\title{Tests whether a column in dataframe, specified by string, is numeric}
\usage{
.is.numeric_col(col, data)
}
\arguments{
\item{col}{Character indicating column name}

\item{data}{Dataframe that contains \code{col}}
}
\description{
Tests whether a column in dataframe, specified by string, is numeric
}
\keyword{internal}
