% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{.round_half_to_even}
\alias{.round_half_to_even}
\title{Round all numeric columns in dataframe}
\source{
https://stackoverflow.com/questions/9063889/how-to-round-a-data-frame-in-r-that-contains-some-character-variables
}
\usage{
.round_half_to_even(x, digits)
}
\arguments{
\item{digits}{Number of digits, defaults to 2}

\item{df}{Dataframe to be rounded}
}
\description{
Rounds all numeric columns in dataframe, using "round half to even" (aka banker's rounding)
}
