% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{.wtd_cor_test_lm}
\alias{.wtd_cor_test_lm}
\title{Calculate the correlation, based on weighted scaling and
listwise deletion \emph{before} scaling}
\usage{
.wtd_cor_test_lm(x, y, wt, ...)
}
\description{
Calculate the correlation, based on weighted scaling and
listwise deletion \emph{before} scaling
}
