% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_quickpred_extension.R
\name{dummy_code}
\alias{dummy_code}
\title{Dummy code variable}
\source{
Simplified from psych::dummy.code, written by William Revelle
}
\usage{
dummy_code(x)
}
\arguments{
\item{x}{A vector to be transformed into dummy codes.}
}
\description{
Simplified from psych::dummy.code and
changed to return data.frame
}
