% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dupl_items}
\alias{dupl_items}
\title{Get duplicated items from a vector}
\usage{
dupl_items(x)
}
\arguments{
\item{x}{A numeric or character vector from which to identify duplicate elements.}
}
\value{
A vector containing unique elements that are duplicated in the original vector.
Returns an empty vector if there are no duplicates.
}
\description{
This function takes a numeric or character vector and returns a vector of the
unique elements that appear more than once - so a simple shortcut for
\code{unique(x[duplicated(x)])} that is particularly friendly in pipes
}
\examples{
dupl_items(c(1, 2, 3, 3, 4, 5, 5, 5))
# Output: 3 5
dupl_items(c(1, 2, 3, 4))
# Output: numeric(0)

}
