% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ess_health}
\alias{ess_health}
\title{Sample data: health data from the European Social Survey Wave 7}
\format{
A dataframe with 7,226 rows and 23 columns
\describe{
\item{cntry}{Country}
\item{gndr}{Gender (1 = male, 2 = female)}
\item{agea}{Age}
\item{eisced}{Education}
\item{pweight}{Population weight - adjust for different population sizes between countries}
\item{pspwght}{Post-stratification weight - adjusts for bias in national samples
due to sampling error and non-response (considering gender, age, education and region)
AND for unequal inclusion probabilities due to survey design}
\item{health}{Overall health, from 1 = very good to 5 = very bad}
\item{height}{Height (cm)}
\item{weight}{Weight (kg)}
\item{icbrnct}{Respondent born in country, 1 = yes, 2 = no}
\item{etfruit}{How often eat fruit, excluding drinking juice, from 1 = three
times or more per day to 7 = never}
\item{eatveg }{How often eat vegetables or salad, excluding potatoes,  from
1 = three times or more per day to 7 = never}
\item{dosprt }{Do sports or other physical activity, how many of last 7 days}
\item{cgtsmke}{Cigarettes smoking behaviour, from 1 = I smoke daily to 5 = I have never smoked}
\item{alcfreq}{How often drink alcohol, from 1 = every day to 5  = never}
\item{fltdpr}{Felt depressed in past week, from 1 = none/almost none of the
time to 4 = all/almost all of the time (same for next 7 variables)}
\item{flteeff}{Felt everything did as effort}
\item{slprl}{Sleep was restless}
\item{wrhpp}{Were happy}
\item{fltlnl}{Felt lonely}
\item{enjlf}{Enjoyed life}
\item{fltsd}{Felt sad}
\item{cldgng}{Could not get going}
}
}
\description{
This dataset contains data from the ESS Wave 7 (2014) for Germany,
the UK and France that describes respondents physical and mental
health as well as specific health behaviours. It can be used to
demonstrate most functions of this package.
}
\references{
\url{https://www.europeansocialsurvey.org/data/download.html?r=7}
}
\author{
ESS-ERIC, selected by package author
}
\keyword{data}
