% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.heritability.R
\name{estimate.heritability}
\alias{estimate.heritability}
\title{Estimates ACE, AE, CE, and E heritability models}
\usage{
estimate.heritability(
  myformula,
  data,
  DZ = "DZ",
  zyg = "ZYGOSITY",
  fid = "FID"
)
}
\arguments{
\item{myformula}{the model to be fit (trait ~ 1 + covariate)}

\item{data}{input data frame}

\item{DZ}{how dyzigotic twins are codified}

\item{zyg}{name of the column describing zygosities}

\item{fid}{name of the column describing family ID}
}
\value{
a list including: 
, ace, the results of the ACE model
, ae, the results of the AE model
, ce, the results of the CE model
, e, the results of the E model
, aics, AICs of all fitted models
, best.m, label for best model (ACE, AE, CE, E)
, best model, the results of the best model
}
\description{
Estimate the contribution of additive genetic (A), shared (C) and individual-specific 
environment (E) effects on traits variations, using a classical twin design study. 
It fits ACE, and the most parsimonious AE, CE e E model, which are then compared using 
the Akaike’s information criterion (AIC).
}
\author{
Alessia Visconti
}
