% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_meff.R
\name{estimate.meff}
\alias{estimate.meff}
\title{Estimates the effective number of tests}
\usage{
estimate.meff(data, method = c("Cheverud", "Li"))
}
\arguments{
\item{data}{data matrix}

\item{method}{which methods should be used}
}
\value{
effective number of tests
}
\description{
Estimates the effective number of tests in correlated datasets using two different 
approaches (Li's is to be preferred)
Cheverud JM (2001). A simple correction for multiple comparisons in interval mapping
genome scans. Heredity 87
}
\examples{
data <- matrix(data = rnorm(200), nrow = 10, ncol = 20)
estimate.meff(data, "Li")
}
\author{
Jonas Zierer
}
