% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format confidence interval based on the bounds}
\usage{
fmt_ci(lower, upper, digits = 2, drop_0 = FALSE)
}
\arguments{
\item{lower}{Lower bound(s) of confidence interval(s). Numeric, or a vector of numbers}

\item{upper}{Lower bound(s) of confidence interval(s). Numeric, or a vector of numbers}

\item{digits}{Number of significant digits, defaults to 2}

\item{drop_0}{Logical. Should leading 0 be dropped, e.g., when reporting correlation coefficients.
Note that this only works when all values are between -1 and 1 (inclusive)}
}
\value{
A character vector of formatted confidence intervals, e.g. \code{"[1.23, 2.68]"},
the same length as \code{lower}/\code{upper}. \code{NA} is returned where either bound is \code{NA}.
}
\description{
Constructs a confidence intervals from upper and lower bounds,
placing them in between square brackets
}
\examples{
fmt_ci(1.23, 2.68)
fmt_ci(c(0.12, 0.45), c(0.34, 0.67), drop_0 = TRUE)
}
