% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_p}
\alias{fmt_p}
\title{Format p-value in line with APA standard (no leading 0)}
\usage{
fmt_p(p_value, digits = 3)
}
\arguments{
\item{p_value}{Numeric, or a vector of numbers}

\item{digits}{Number of significant digits, defaults to 3}
}
\description{
Formats p-value in line with APA standard, returning it without leading 0 and
as < .001 when it is that small.
}
