% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.inflation.factor.R
\name{genomic.inflation.factor}
\alias{genomic.inflation.factor}
\title{Evalutes the genomic inflation factor (lambda)}
\usage{
genomic.inflation.factor(values, type = "pvalue")
}
\arguments{
\item{values}{the statistic distribution}

\item{type}{statistic distribution (allowed: zscore, chisq, pvalue -- default: pvalue)}
}
\value{
the "inflation factor" lambda
}
\description{
Evalutes the genomic inflation factor (lambda) for genome-wide association studies. 
In details, lambda is the observed median value of the chi-squared statistic for the 
null markers divided by the expected median  value of the chi-squared statistic 
(approximately 0.456 for 1 degree of freedom tests).
}
\examples{
pvalues <- abs(rnorm(1000, mean=0.001, sd=0.1))
genomic.inflation.factor(pvalues, "pvalue")
}
\author{
Alessia Visconti
}
