% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age.R
\name{get.age}
\alias{get.age}
\title{Evaluates age difference between dates}
\usage{
get.age(born, visit, format.born = "\%m/\%d/\%Y",
  format.visit = "\%m/\%d/\%Y")
}
\arguments{
\item{born}{date of birth (character)}

\item{visit}{date of visit (character)}

\item{format.born}{date of birth format}

\item{format.visit}{date of visit format}
}
\value{
age at visit
}
\description{
Given the date of birth and the date of visit, along with their format, it evaluates the 
subject's age at visit
}
\examples{
get.age("01/01/1980", "31/12/2019", "\%d/\%m/\%Y", "\%d/\%m/\%Y")
}
\author{
Alessia Visconti
}
