% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeframe.R
\name{get.timeframe}
\alias{get.timeframe}
\title{Evaluates day, week, month, and age differences between dates}
\usage{
get.timeframe(date1, date2, format1 = "\%m/\%d/\%Y",
  format2 = "\%m/\%d/\%Y")
}
\arguments{
\item{date1}{first date (string)}

\item{date2}{second date (string)}

\item{format1}{format of date1}

\item{format2}{format of date2}
}
\value{
difference between the dates
}
\description{
Given two dates, it returns the differences in days, weeks, months, and years.
Each month is estimated to be 30 day long.
}
\examples{
get.timeframe("01/01/1980", "31/12/2019", "\%d/\%m/\%Y", "\%d/\%m/\%Y")
}
\author{
Alessia Visconti
}
