% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_coef_rename_tribble}
\alias{get_coef_rename_tribble}
\title{Get code to rename model coefficients in summary tables}
\usage{
get_coef_rename_tribble(mod, show = TRUE)
}
\arguments{
\item{mod}{A model (e.g., returned from lm). coef() is called to extract the coefficient names.}

\item{show}{Logical - should the output be printed to the console. In any case, it is returned invisibly}
}
\value{
Code to be edited and passed to tibble::tribble() to create the coef_renames argument for
\code{\link{report_lm_with_std}} and similar functions. 'New' coefficient names are minimally
transformed, but are primarily intended to be placeholders.
}
\description{
Usually, the coefficients in a statistical model such as those returned from `lm()`
need to be renamed for reporting, particularly when dummy/indicator variables are involved.
This function returns the code to set up a rename tribble that can then be adjusted
and passed, for instance, to the \code{\link{report_lm_with_std}} function.
}
\examples{
mod <- lm(mpg ~ wt, mtcars)
get_coef_rename_tribble(mod)
}
