% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{get_pairwise_letters}
\alias{get_pairwise_letters}
\title{Get letters to indicate results of multiple comparisons/post-hoc tests}
\source{
Algorithm based on finding maximal cliques of non-significant comparisons.
}
\usage{
get_pairwise_letters(tests, alpha_level = 0.05)
}
\arguments{
\item{tests}{Either a tibble with the result of comparisons, including x and y
(the levels/groups that were compared) and p_value for the comparison or an object
of class pairwise.htest, for example returned from pairwise.t.test()}

\item{alpha_level}{The level of significance for the test}
}
\value{
A tibble with the group level, a combined letters string, and
individual columns for each letter assigned. The tibble is sorted by level.
}
\description{
This takes the results of multiple comparisons and returns a set of letters
that can be used to indicate which differences are significant. The difference
between levels that are assigned the same letter are \emph{not} statistically different.
}
\examples{
# Works with pairwise.htest objects
library(tibble)
library(dplyr)

pwtt <- pairwise.t.test(airquality$Ozone, airquality$Month)
get_pairwise_letters(pwtt)

}
