% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_rename_tribbles}
\alias{get_rename_tribbles}
\title{Get code to generate tibbles to rename categorical variables and their levels}
\usage{
get_rename_tribbles(
  df,
  ...,
  show = TRUE,
  which = c("both", "vars", "levels"),
  max_levels = 20
)
}
\arguments{
\item{df}{A dataframe that contains the variables - only used to extract their possible levels.}

\item{...}{The variables to be included in the rename tribbles.}

\item{show}{Logical - should the output be printed to the console. In any case, it is returned invisibly}

\item{which}{Should tribble code be generated for variables (\code{"vars"}), levels (\code{"levels"}) or both (\code{"both"}) (default)}

\item{max_levels}{The maximum number of levels before a variable is dropped from the levels_tribble. Defaults to 20}
}
\value{
Code to be edited and passed to tibble::tribble() to create var_names and level_names arguments for
\code{\link{rename_cat_variables}} and \code{\link{cat_var_table}}
}
\description{
Renaming categorical variables and their levels, for instance for summary tables, can be fiddly. This
function generates code in which only the new names need to be modified, and which can then be passed
to either \code{\link{rename_cat_variables}} or directly to \code{\link{cat_var_table}}
}
\details{
Only categorical variables should be passed to the function if code for levels is
requested. If a variable has more than 20 distinct values, it is dropped from the levels-tribble-code,
unless the max_levels argument is increased.
}
\examples{
get_rename_tribbles(iris)

}
