% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{ggsave_show}
\alias{ggsave_show}
\title{Save ggplot-graph and show in folder}
\source{
https://stackoverflow.com/a/12135823/10581449
}
\usage{
ggsave_show(filename = NULL, ..., device = NULL, units = "cm")
}
\arguments{
\item{filename}{File name with path. If not provided, only a temporary file is saved}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{plot}}{Plot to save, defaults to last plot displayed.}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width,height}}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
    \item{\code{create.dir}}{Whether to create new directories if a non-existing
directory is specified in the \code{filename} or \code{path} (\code{TRUE}) or return an
error (\code{FALSE}, default). If \code{FALSE} and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}
  }}

\item{device}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}

\item{units}{Unit for width and height, if provided. Defaults to "cm", can also be "in" or "mm"}
}
\description{
This wraps \code{ggsave} and opens the folder where the graph was saved in a Shell.
From there, it can easily be dragged and dropped into the application where you want to use it.
It also changes the default units from in to cm, and defaults to saving temporary png files.
}
\examples{
\dontrun{
ggplot(mtcars, aes(mpg)) + geom_histogram()
ggsave_show(here::here("mtcars.pdf"))
}
}
