% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tables.R
\name{glance.mira}
\alias{glance.mira}
\title{Glance a multiple imputation \code{mice} pooled object}
\usage{
\method{glance}{mira}(x, adj_R2 = TRUE, ...)
}
\arguments{
\item{x}{An object with multiply-imputed models from \code{mice} (class: \code{mira})}

\item{adj_R2}{Should an adjusted R2 be returned as well? Only applicable if x contains \code{lm}-models}

\item{...}{extra arguments (not used)}
}
\value{
a tibble with one row
}
\description{
Note that the \code{mice} authors prefer to tidy \code{mipo} rather than \code{mira} objects and have now included \code{tidy.mipo} and \code{glance.mipo} into their package. The \code{mira} functions here are mostly retained for compatibility with my earlier code.
}
\note{
If x contains \code{lm} models, R2 is included in the output
}
\examples{
\dontrun{
library(mice)
data <- airquality
data[4:10, 3] <- rep(NA, 7)
data[1:5, 4] <- NA
tmp <- mice(data, m = 5, seed = 500, printFlag = FALSE)
mod <- with(tmp, lm(Ozone ~ Solar.R + Wind))
glance(mod)
}
}
