% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{gt_add_plots}
\alias{gt_add_plots}
\title{Add plots into gt table column}
\usage{
gt_add_plots(gt_table, plots, col_index)
}
\arguments{
\item{gt_table}{A gt table to add the plots into}

\item{plots}{A list of ggplot2 plots, typically with the same length as the number of rows in gt_table}

\item{col_index}{The index of the column in gt_table that is to be overwritten with the plots}
}
\description{
This function takes a list of ggplot2 plots and adds them into a gt table column.
}
\examples{
\dontrun{
var_names <- c(wt = "Weight", am = "Transmission", mpg = "Consumption (mpg)", gear = "Gears")
cor_table <- cor_matrix(mtcars, var_names) \%>\%
  report_cor_table(extras = tibble::tibble(Distributions = c(1:length(var_names))))
large_text <- ggplot2::theme(axis.text.x = ggplot2::element_text(size=40))
distr_plots <- plot_distributions(mtcars, var_names, plot_theme = large_text)
gt_add_plots(cor_table, distr_plots, 3)
}
}
