% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{gt_apa_style}
\alias{gt_apa_style}
\title{Helper function to style gt-table in APA style}
\source{
Created by Philip Parker, https://gist.github.com/pdparker/1b61b6d36d09cb295bf286a931990159. Slightly expanded here.
}
\usage{
gt_apa_style(gt_table, fmt_labels_md = TRUE)
}
\arguments{
\item{gt_table}{A gt-table}

\item{fmt_labels_md}{Should row and column labels be formatted with markdown/HTML (Defaults to TRUE)}
}
\description{
This function takes a `gt` table object and changes font-type, borders etc
to align with APA style.
}
