% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_GWAS_Catalog.R
\name{GWAS.catalog.SNP}
\alias{GWAS.catalog.SNP}
\title{Fetches associations in the GWAS Catalog involving the gived SNP}
\usage{
GWAS.catalog.SNP(rs)
}
\arguments{
\item{rs}{reference SNP IDs for the variant}
}
\value{
a data frame listing the associations recorded in the GWAS catalog
}
\description{
Uses GWAS Catalog Access via API (\url{https://www.ebi.ac.uk/gwas/rest/docs/api}).
to fetch GWAS associations. If no associations is found, NULL is returned.
}
\examples{
GWAS.catalog.SNP("rs7681615")
GWAS.catalog.SNP("rs4693052")
}
\author{
Alessia Visconti
}
