% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_normal.R
\name{inverse.normal}
\alias{inverse.normal}
\title{Performs the inverse normal transformation}
\usage{
inverse.normal(v)
}
\arguments{
\item{v}{vector of values}
}
\value{
a vector of transformed vectors
}
\description{
Performs the inverse normal transformation
}
\examples{
x <- rexp(100)
inverse.normal(x)
}
\author{
Alessia Visconti
}
