% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlink_query.R
\name{LDpair}
\alias{LDpair}
\title{Fetches LD (r2) between two SNPs}
\usage{
LDpair(rs1, rs2, pop = "GBR", LDlinktoken)
}
\arguments{
\item{rs1}{reference SNP IDs for the first variant}

\item{rs2}{reference SNP IDs for the second variant}

\item{pop}{population (multiple populations should be formatted as pop1\%2pop2\%2pop3)}

\item{LDlinktoken}{personal token (request via \url{https://ldlink.nci.nih.gov/?tab=apiaccess})}
}
\value{
LD between the two given SNPs in the given population, or NA if one or both of the SNPs is not found
}
\description{
Uses LDlink Programmatic Access via API (\url{https://ldlink.nci.nih.gov/?tab=ldpair}).
LDlink allows extracting LD between two given SNPs in a given population.
According to the LDLink CoC, this script should be run sequentially.
}
\examples{
LDpair("rs123", "rs456", "GBR", "n0tw0rk1ng")
}
\author{
Alessia Visconti
}
