% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.ratio.test.R
\name{likelihood.ratio.test}
\alias{likelihood.ratio.test}
\title{Evaluates likelihood ratio test
 
Compares the goodness of fit of a null model (LO) against an alternative model (L1)}
\usage{
likelihood.ratio.test(L0, L1)
}
\arguments{
\item{L0}{null model likelihood}

\item{L1}{alternative model likelihood}
}
\value{
p value assessing the goodness of fit
}
\description{
Evaluates likelihood ratio test
 
Compares the goodness of fit of a null model (LO) against an alternative model (L1)
}
\examples{
data(mtcars)
model0 <- lm(qsec ~ 1, data=mtcars)
model1 <- lm(qsec ~ cyl, data=mtcars)
likelihood.ratio.test(logLik(model0), logLik(model1))
}
\author{
Alessia Visconti
}
