% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{line_to_vector}
\alias{line_to_vector}
\alias{l2v}
\title{Turn a Line of Items into a Vector or c() Code}
\usage{
line_to_vector(
  x = NULL,
  strings = NULL,
  separators = c("top-level", "all"),
  to_clip = interactive(),
  return = c("code", "vector"),
  keep_blank_as_na = FALSE
)

l2v(
  x = NULL,
  strings = NULL,
  separators = c("top-level", "all"),
  to_clip = interactive(),
  return = c("code", "vector"),
  keep_blank_as_na = FALSE
)
}
\arguments{
\item{x}{Character string of desired vector items, separated by spaces, tabs,
or line breaks. If \code{NULL}, the function will attempt to read from the clipboard.}

\item{strings}{Logical or \code{NULL}. Determines whether vector items should be
treated as strings (i.e., quoted). Defaults to \code{TRUE} unless all items are numeric.}

\item{separators}{Character string indicating which separator should be used
to split \code{x}. Defaults to the highest-level separator found in \code{x} (i.e.,
newlines if present). Can also be \code{"all"} to split by all supported separators
(i.e. spaces: \code{" "}, commas: \code{","}, tabs: \code{"\\t"}, newlines: \code{"\\n"}).}

\item{to_clip}{Logical. Indicates whether the generated code for the vector
should be copied to the clipboard. Defaults to \code{TRUE} in interactive sessions.
Requires the \code{clipr} package to be installed.}

\item{keep_blank_as_na}{Logical. If \code{TRUE}, empty values are kept as \code{NA}.
If \code{FALSE} (default), empty values are omitted from the resulting vector.}

\item{return_type}{Character string specifying the type of return value.
Choose \code{"code"} to return the \code{c()} command as a string or \code{"vector"} to
return the actual R vector. Defaults to \code{"code"}. Abbreviated values like
\code{"c"} or \code{"v"} are accepted.}
}
\value{
A character string representing the \code{c()} command or the actual R
vector, depending on the \code{return_type} parameter.
}
\description{
This function takes a line of items separated by spaces, tabs, or newlines and
returns either a \code{c()} vector command (with items quoted or not) or the vector
created by that command. By default, the line is read from the clipboard and
a character vector is returned. The function uses only the highest-level
separator present in \code{x} unless specified otherwise. This functionality is
particularly useful for copying values from spreadsheet programs like Excel
(or from R console output) and converting them into R vectors seamlessly.
}
\details{
If \code{x} is not split by spaces, \code{stringr::str_trim()} is used to trim
whitespace from the start and end of each element of the vector.
}
\examples{

line_to_vector("a b c", strings = TRUE)
# c("a", "b", "c")


line_to_vector("1 2 3", return_type = "vector")
# [1] 1 2 3

# Can abbreviate return argument
# [1] "Friday"   "Saturday" "Sunday"

}
