% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scales.R
\name{make_scales}
\alias{make_scales}
\title{Creates multiple scales by calculating item means and returns descriptives}
\usage{
make_scales(
  df,
  items,
  reversed = NULL,
  two_items_reliability = c("spearman_brown", "cronbachs_alpha", "r"),
  ...
)
}
\arguments{
\item{df}{A dataframe}

\item{items}{A named list of characters vectors. Names are the scale names,
each vector contains the items for that scale (variables in df)}

\item{reversed}{A named list of characters vectors. Names are the scale names,
each vector contains the items to be reverse-coded for that scale}

\item{two_items_reliability}{How should the reliability of two-item scales be
reported? "spearman_brown" is the recommended default, but "cronbachs_alpha"
and Pearson's "r" are also supported.}

\item{...}{
  Arguments passed on to \code{\link[=make_scale]{make_scale}}
  \describe{
    \item{\code{print_hist}}{Logical. Should histograms for items and resulting scale be printed?}
    \item{\code{print_desc}}{Logical. Should descriptives for scales be printed?}
  }}
}
\value{
A list of two dataframes: scale values (`scores`) and
descriptive statistics for each scale (`descriptives`)
}
\description{
This function creates multiple scales, returns descriptives and supports
reverse-coding of items.
}
