% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{na_share}
\alias{na_share}
\title{Calculate share of NA-values in vector}
\usage{
na_share(x, round = 2)
}
\arguments{
\item{x}{Vector}

\item{round}{Number of digits to round result to}
}
\value{
Numeric. The share of \code{NA} values in \code{x}, rounded to \code{round} digits.
}
\description{
Returns share of NA values in vector by wrapping \code{sum(is.na(x)) / length(x)}
and rounding the result
}
\examples{
x <- c(NA, 1, 2)
na_share(x)
}
