% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{named_list}
\alias{named_list}
\title{Create a list with items named based on the names of variables passed to the function}
\usage{
named_list(...)
}
\arguments{
\item{...}{One or more variables to be included into a list. Note that you can pass values as
well, but they are then used as both the name and the value of the respective list item.}
}
\value{
A named list, with names taken from the variable names in the input
}
\description{
Not infrequently, you might find yourself writing things like \code{list(var1 = var1, var2 = var2)}.
This function makes this easier by creating a list with items named based on the variables
passed into it.
}
\examples{
name <- "Paul"
age <- 10
named_list(name, age)

# Note that this can be combined with named arguments
named_list(name, age, place = "Berlin")

}
\keyword{internal}
