% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{pairwise_t_test_mi}
\alias{pairwise_t_test_mi}
\title{Pairwise t-tests() on multiply-imputed data (accepts survey weights)}
\usage{
pairwise_t_test_mi(
  mi_list,
  dv,
  groups,
  weights = NULL,
  p.adjust.method = p.adjust.methods
)
}
\arguments{
\item{mi_list}{A list of dataframes, each consisting of a multiply imputed dataset}

\item{dv}{The dependent variable for the t.test (must be in mi_list)}

\item{groups}{The grouping variable (each distinct value will be treated as a level)}

\item{weights}{The variable containing survey weights, must be in mi_list}

\item{p.adjust.method}{The method to adjust p-values for multiple comparison (see \code{\link[stats]{p.adjust}})}
}
\value{
A tibble containing the results of the t-tests with one test per row
}
\description{
This runs pairwise t-tests (based on lm(), therefore assuming equal variances)
on multiple imputations, with the ability to take into account survey weights.
}
