% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_}
\alias{paste_}
\title{Paste arguments while ignoring NAs}
\usage{
paste_(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{The arguments to be concatenated. These can be vectors.}

\item{sep}{The string that separates the concatenated elements. Defaults to a single space.}

\item{collapse}{An optional string to separate the results. NULL by default, which means no separation.}
}
\value{
A character vector of concatenated strings.
}
\description{
This function behaves like base::paste(), but replaces any NA elements with an empty string.
It also supports vector recycling and accepts the 'sep' and 'collapse' arguments.
}
\examples{
paste_("hello", NA, "you") 
# returns "hello you"

paste_(c("hello", "world"), c(NA, "everyone"), c("you", NA)) 
# returns c("hello you", "world everyone")

}
