% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_quickpred_extension.R
\name{quickpred_ext}
\alias{quickpred_ext}
\title{Create mice predictorMatrix including unordered factors (extends mice::quickpred)}
\source{
This function is based on the mice::quickpred function as available in mice v 3.11.4, written by Stef van Buuren. The code for dummy coding factor variables is based on psych::dummy.code, written by William Revelle
}
\usage{
quickpred_ext(
  data,
  mincor = 0.1,
  minpuc = 0,
  include = "",
  exclude = "",
  method = "pearson"
)
}
\arguments{
\item{data}{Matrix or data frame with incomplete data.}

\item{mincor}{A scalar, numeric vector (of size \code{ncol(data))} or numeric
matrix (square, of size \code{ncol(data)} specifying the minimum
threshold(s) against which the absolute correlation in the data is compared.}

\item{minpuc}{A scalar, vector (of size \code{ncol(data))} or matrix (square,
of size \code{ncol(data)} specifying the minimum threshold(s) for the
proportion of usable cases.}

\item{include}{A string or a vector of strings containing one or more
variable names from \code{names(data)}. Variables specified are always
included as a predictor.}

\item{exclude}{A string or a vector of strings containing one or more
variable names from \code{names(data)}. Variables specified are always
excluded as a predictor.}

\item{method}{A string specifying the type of correlation. Use
\code{'pearson'} (default), \code{'kendall'} or \code{'spearman'}. Can be
abbreviated.}
}
\value{
A square binary matrix of size \code{ncol(data)}.
}
\description{
\code{\link[mice]{quickpred}} creates a predictor matrix that suggests which variables should
be used for multiple imputation. However, it does not yield meaningful information
for (unordered) factors. This function returns the same as mice::quickpred for
numeric and logical variables and binary factors, but excludes character
vectors and tests the predictive power of each level of factors with more
than two levels separately (through dummy coding). If the correlation exceeds `mincor` for at least one
level, the factor variable is used as a predictor.
}
\seealso{
\code{\link[mice]{quickpred}}
}
