% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tables.R, R/timesaveR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glance}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{glance}}}

  \item{generics}{\code{\link[generics]{tidy}}}
}}

