% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_anova}
\alias{report_anova}
\title{Report (model comparison) ANOVA tests}
\usage{
report_anova(x)
}
\arguments{
\item{x}{An object returned from an anova() function (or car::linearHypothesis)}
}
\value{
A character string reporting the anova in APA style
}
\description{
This is under development - contributions to support more types of ANOVAs are most welcome.
}
\examples{
mod1 <- lm(mpg ~ wt + am, mtcars)
mod2 <- lm(mpg ~ wt * am, mtcars)
report_anova(anova(mod1, mod2))
}
