% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{report_cor_table}
\alias{report_cor_table}
\title{Create a correlation table with summary statistics in APA style}
\source{
Based on the apaTables \code{apa.cor.table()} function, but adapted to
accept weighted correlation matrices and work with the \code{gt} package instead. Code
for calculation of confidence intervals adapted from
https://medium.com/@shandou/how-to-compute-confidence-interval-for-pearsons-r-a-brief-guide-951445b9cb2d`
}
\usage{
report_cor_table(
  cor_matrix,
  ci_type = deprecated(),
  ci_width = deprecated(),
  ci = deprecated(),
  n = deprecated(),
  add_distributions = FALSE,
  data = NULL,
  filename = NULL,
  notes = list(NULL),
  stars = NULL,
  add_title = FALSE,
  extras = NULL,
  apa_style = TRUE,
  ...
)
}
\arguments{
\item{cor_matrix}{A correlation matrix, for example returned from
\code{\link[=cor_matrix]{cor_matrix()}}, \code{\link[=svy_cor_matrix]{svy_cor_matrix()}}, or \code{\link[=cor_matrix_mi]{cor_matrix_mi()}}.}

\item{ci_type}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} Recalculating CIs in this
function has been removed. Please generate CIs in \code{cor_matrix()}.}

\item{ci_width}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} Use \code{conf_level} in \code{cor_matrix()} instead.}

\item{ci}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} Superseded by functionality in \code{cor_matrix()}.}

\item{n}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} No longer used.}

\item{add_distributions}{Logical. Add graphs showing variable distributions?
Works with both regular data frames and survey design objects.}

\item{data}{Original data, only needed if \code{add_distributions = TRUE}. Can be a
regular data frame or a survey design object. For survey data, distributions
will be properly weighted using \code{svyhist()} and \code{svysmooth()}.}

\item{filename}{The file name to create on disk. Include '.html' extension to
best preserve formatting (see \code{gt::gtsave} for details).}

\item{notes}{List of additional notes to show under the table.}

\item{stars}{A character vector to change the significance symbols (see details below)}

\item{add_title}{Should title be added to table? Set to TRUE for default
title or provide a character string for custom title.}

\item{extras}{Tibble of additional columns to be added after the descriptives column -
needs to be sorted in the same order as the \code{desc} element in the \code{cor_matrix} unless
there is a \code{row_names} column. If there is, this will be used to match it to the \code{desc} rows.}

\item{apa_style}{Logical, should APA-style formatting be applied.}

\item{...}{
  Arguments passed on to \code{\link{plot_distributions}}
  \describe{
    \item{\code{plot_type}}{Type of plot that should be produced - \code{histogram} or \code{density} plot. If \code{auto},
histograms are produced for variables that take fewer than 10 unique values, density plots for others. If a number is provided,
that number is used as the maximum number of unique values for which a histogram is used.}
    \item{\code{hist_align_y}}{Should histograms use the same y-axis, so that bin heights are comparable? Defaults to FALSE}
    \item{\code{plot_theme}}{Additional theme_ commands to be added to each plot}
  }}
}
\value{
A \code{gt} table that can be further formatted with \code{gt}-functions.
}
\description{
This function creates (and optionally saves) a correlation table with
summary statistics. It accepts correlation matrices from various functions
in this package as its first argument.
}
\examples{
# Basic correlation table
cor_matrix(iris, var_names = c(Sepal.Length = "Sepal Length",
                                Sepal.Width = "Sepal Width")) \%>\%
  report_cor_table()

# With distributions
cor_matrix(iris) \%>\%
  report_cor_table(add_distributions = TRUE, data = iris,
                   add_title = "Iris correlations and distributions")

# Using ESS health data
cor_matrix(ess_health, var_names = c(health = "Health",
                                      dosprt = "Physical activity",
                                      etfruit = "Fruit consumption")) \%>\%
  report_cor_table()

# With survey data and distributions
\dontrun{
library(survey)
library(srvyr)
data(api)
dstrat <- apistrat \%>\% as_survey_design(1, strata = stype, fpc = fpc, weight = pw)
svy_cor_matrix(dstrat, var_names = c(enroll = "Enrollment",
                                      api00 = "API 2000",
                                      api99 = "API 1999")) \%>\%
  report_cor_table(add_distributions = TRUE, data = dstrat)
}

}
