% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scale_weighted}
\alias{scale_weighted}
\title{Create a Z-score (standardize) for a vector using weights.}
\usage{
scale_weighted(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights. Must be the same length as x.}

\item{na.rm}{If TRUE, remove cases where x or w is NA.}
}
\value{
A numeric vector of the same length as x with weighted Z-scores.
}
\description{
Create a Z-score (standardize) for a vector using weights.
}
\keyword{internal}
