% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\encoding{UTF-8}
\name{sigstars}
\alias{sigstars}
\title{Significance stars for p-values}
\source{
Adapted from
 http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\usage{
sigstars(p, stars = NULL, pad_html = FALSE, ns = FALSE, return_NAs = FALSE)
}
\arguments{
\item{p}{A \emph{p}-value or (more commonly) a vector of \emph{p}-values}

\item{stars}{A character vector to change the significance symbols (see details in `sigstars`)}

\item{pad_html}{Should all results be padded right to the same width with HTML non-breaking spaces?}

\item{ns}{Logical. Should non-significant values be highlighted as "ns"?}

\item{return_NAs}{Logical. Should NAs be returned? If not, empty strings are returned instead.}
}
\value{
A character vector of significance stars for each \emph{p}-value,
each padded with spaces to be four characters long
}
\description{
Function returns significance stars for \emph{p}-values, most likely for use
in tables that report the results of multiple statistical tests. An empty
string is returned for NAs, unless that behaviour is overwritten.
}
\details{
Symbols and thresholds are *** \emph{p} < .001,  ** \emph{p} < .01, * \emph{p}
 < .05 and † \emph{p} < .1. The symbols can be changed by passing a named character vector sorted
 descendingly to the \code{stars} argument. For the default, the argument would be
\code{stars <- c(`&dagger;` = .1, `*` = 0.05, `**` = 0.01, `***` = 0.001)}
}
