% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{summary.rN_lm}
\alias{summary.rN_lm}
\title{Summary of an lm object created with run_lm() wrapper}
\usage{
\method{summary}{rN_lm}(object, ...)
}
\arguments{
\item{object}{A model with class `rN_lm`}

\item{...}{
  Arguments passed on to \code{\link[stats:summary.lm]{stats::summary.lm}}
  \describe{
    \item{\code{correlation}}{logical; if \code{TRUE}, the correlation matrix of
    the estimated parameters is returned and printed.}
    \item{\code{symbolic.cor}}{logical. If \code{TRUE}, print the correlations in
    a symbolic form (see \code{\link[stats]{symnum}}) rather than as numbers.}
  }}
}
\description{
Using \code{\link{run_lm}} creates a very unwieldy call to lm(). This
function replaces it by a more legible call in the `summary()`-output
}
