% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{svy_pairwise.t.test}
\alias{svy_pairwise.t.test}
\title{Pairwise t.tests with effect sizes and survey weights}
\usage{
svy_pairwise.t.test(df, dv, iv, cats, ...)
}
\arguments{
\item{df}{A dataframe}

\item{dv}{Character. Name of the dependent variable for the t.test (numeric)}

\item{iv}{Character. Name of the grouping variable for the t.test (factor)}

\item{cats}{Character vector of factor levels to be included in the
pairwise tests. If set to NULL, all levels are used.}

\item{...}{
  Arguments passed on to \code{\link[=svy_cohen_d_pair]{svy_cohen_d_pair}}
  \describe{
    \item{\code{ttest}}{Logical. Should t.test be run and displayed? Otherwise, only
Cohen's d is calculated. Defaults to TRUE.}
  }}
}
\value{
Invisibly returns a names lists of lists including the t.test()
  output and Cohen's D for each pair
}
\description{
This function calculates a t.test() for any pair of levels in a
\code{srvyr} survey object. It does currently not do any p-value adjustment
for multiple comparisons, and print rather than returns the results.
}
