% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{t_test_mi}
\alias{t_test_mi}
\title{t-test() on multiply-imputed data (accepts survey weights)}
\usage{
t_test_mi(mi_list, dv, groups, weights = NULL)
}
\arguments{
\item{mi_list}{A list of dataframes, each consisting of a multiply imputed dataset}

\item{dv}{The dependent variable for the t.test (must be in mi_list)}

\item{groups}{The grouping variable (must have only two values, be in mi_list)}

\item{weights}{The variable containing survey weights, must be in mi_list}
}
\value{
A one-row tibble containing the result of the t-test
}
\description{
This runs t-test (based on lm(), therefore assuming equal variances) on multiple imputations,
with the ability to take into account survey weights.
}
\examples{
#Create list with imputed data
library(mice)
imp <- mice(nhanes2)
imp_list <- complete(imp, action="long") \%>\%
   dplyr::group_split(.imp)

t_test_mi(imp_list, bmi, hyp)

}
