% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidy.exp}
\alias{tidy.exp}
\title{Tidy function to exponentiate coefficients}
\usage{
tidy.exp(x, ...)
}
\arguments{
\item{x}{An object, usually containing a logistic regression model. Should
have the class "exp" as the first of its classes, and then a class that dispatches
it to an appropriate `generics::tidy`` function}

\item{...}{Arguments passed on to the appropriate `tidy` function}
}
\description{
This function calls the tidy method based on the second class of the
object (i.e. after removing the "exp" class that led to it being called),
and then exponentiates the returned estimates and confidence intervals (if any)
in the tibble. This is usually used to turn coefficients of logistic
regression models into Odds Ratios.
}
