% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tables.R
\name{tidy.mira}
\alias{tidy.mira}
\title{Tidy  multiple imputation models created with \code{mice}}
\usage{
\method{tidy}{mira}(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A \code{mira} object containing multiple models based on \code{mice} imputations.}

\item{conf.int}{Logical. Should confidence intervals be returned. Defaults to true.}

\item{conf.level}{Confidence level for intervals. Defaults to .95}

\item{...}{extra arguments (not used)}
}
\description{
Note that the \code{mice} authors prefer to tidy \code{mipo} rather than \code{mira} objects and have now included \code{tidy.mipo} and \code{glance.mipo} into their package. The \code{mira} functions here are mostly retained for compatibility with my earlier code.
}
\note{
Available stats in result:
\itemize{
\item estimate
\item ubar
\item b
\item t
\item dfcom
\item df
\item riv
\item lambda
\item fmi
\item p.value
\item conf.low (if called with conf.int = TRUE)
\item conf.high (if called with conf.int = TRUE)
}
}
