% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{tidy.rN_std}
\alias{tidy.rN_std}
\title{Helper function to enable tidy.lm to be used on lm_std() models}
\usage{
\method{tidy}{rN_std}(x, ...)
}
\arguments{
\item{x}{An object with class rN_std}

\item{...}{arguments passed on to tidy method}
}
\description{
Strips rN_std class and calls tidy() again
}
