% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{tidy.svy_cor_matrix}
\alias{tidy.svy_cor_matrix}
\title{Tidy a survey-weighted correlation matrix}
\usage{
\method{tidy}{svy_cor_matrix}(x, both_directions = TRUE, ...)
}
\arguments{
\item{x}{A \code{svy_cor_marix} object returned from
\code{\link[=svy_cor_matrix]{svy_cor_matrix()}}}

\item{both_directions}{Should both  \code{cor(A, B)}
and \code{cor(B, A)} be returned. Defaults to \code{TRUE}.}

\item{...}{Additional arguments. Not used. Needed to match generic signature
only.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\item{column1}{Name of the first variable}
\item{column2}{Name of the second variable}
\item{estimate}{The estimated value of the correlation}
\item{statistic}{The t-statistic used for significance testing}
\item{p.value}{The two-sided p-value of the correlation}
}
\description{
This function turns the correlation matrix returned by \code{\link[=svy_cor_matrix]{svy_cor_matrix()}}. Note that by default, results for both the \code{cor(A, B)}
and \code{cor(B, A)} are returned, while entries for \verb{A, A}, i.e. the values on the diagonal,
are never included.
}
