% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tables.R
\name{tidy.tsR_std}
\alias{tidy.tsR_std}
\title{Helper function to enable tidiers to be used on standardized models}
\usage{
\method{tidy}{tsR_std}(x, ...)
}
\arguments{
\item{x}{An object with class tsR_std}

\item{...}{arguments passed on to tidy method}
}
\value{
A tidy \code{\link[tibble:tibble]{tibble::tibble()}} of model results, as returned by the
underlying model's \code{tidy()} method (e.g. \code{broom::tidy.lm()}), after
stripping the \code{tsR_std} class.
}
\description{
Strips tsR_std class and calls tidy() again
(for use with lm_std() and polr_std() outputs)
}
\examples{
mod_std <- lm_std(mpg ~ hp + wt, data = mtcars)
tidy(mod_std)
}
