% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesaveR-package.R
\docType{package}
\name{timesaveR}
\alias{timesaveR-package}
\alias{timesaveR}
\title{Functions to Accelerate (Academic) Data Analysis and Reporting}
\description{
Functions and templates in this package facilitate common tasks
in the course of most research projects in social psychology and related
fields. These include creating scales, reporting descriptive statistics,
correlations and distributions, reporting linear regression models with
standardized coefficients and F-change comparisons, as well as plotting
mediation models.
}
\section{Package Options}{

The package recognizes the following option:
\describe{
\item{\code{timesaveR.round_method}}{Controls rounding behavior throughout the
package. Options are \code{"to_even"} (default; banker's rounding) or \code{"default"}
(standard R rounding). Set via \code{options(timesaveR.round_method = "default")}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://lukaswallrich.github.io/timesaveR/}
  \item Report bugs at \url{https://github.com/LukasWallrich/timesaveR/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID}) [copyright holder]
}

}
