% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.chisq.R
\name{to.chisq}
\alias{to.chisq}
\title{Transform statistical distribition to chi-squared values}
\usage{
to.chisq(values, type = c("zscore", "chisq", "pvalue"))
}
\arguments{
\item{values}{a vector of numerical values}

\item{type}{which type of statistic distribution}
}
\value{
chi-squared values
}
\description{
Given a vector of numerical values and which type of statistical
distribition thet are (i.e., chisq, pvalue, zscore) it coverts 
them in chi-squared.
}
\examples{
pvalues <- abs(rnorm(100, mean=0, sd=0.1))
to.chisq(pvalues, "pvalue")
}
\author{
Alessia Visconti
}
