% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{WVS}
\alias{WVS}
\title{Sample data: World Values Survey}
\description{
This dataset contains responses from the World Values Surveys 1995-1997 for 
Australia, Norway, Sweden, and the United States. It is taken from the
\code{carData} package.
}
\details{
#' @format A dataframe with 5,381 rows and 23 columns
\describe{
 \item{poverty}{“Do you think that what the government is doing for people in 
 poverty in this country is about the right amount, too much, or too little?” (ordered): \code{Too Little, About Right, Too Much}}
 \item{religion}{Member of a religion: \code{yes} or \code{no}}
 \item{degree}{Held a university degree: \code{yes} or \code{no}}
 \item{country}{\code{Australia, Norway, Sweden} or \code{USA}}
 \item{age}{in years}
 \item{gender}{\code{male} or \code{female}}
 }
}
\author{
Taken from carData package
}
\keyword{data}
