% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{biomart.SNPid.in.window}
\alias{biomart.SNPid.in.window}
\title{Fetches the reference SNP IDs in the spefified window}
\usage{
biomart.SNPid.in.window(
  chr,
  start,
  end,
  mart,
  with.position = FALSE,
  with.alleles = FALSE
)
}
\arguments{
\item{chr}{chromosome number (1, 2, ...)}

\item{start}{fist position of the window (included)}

\item{end}{last position of the window (included)}

\item{mart}{BioMart database}

\item{with.position}{whether including chromosome name and start and end position (default: false)}

\item{with.alleles}{whether including variant alleles (default: false, if with.position==TRUE is forced to TRUE)}
}
\value{
reference SNP IDs, or reference SNP IDs plus genomic coordinates and alleles
}
\description{
Uses biomaRt function to fetch the reference SNP IDs for those SNPs included in the 
specified window. It uses the given BioMart database.
}
\examples{
mybiomart <- biomart.fetch.SNP.grch37.mart()
biomart.SNPid.in.window(1, 1, 10100, mybiomart)
biomart.SNPid.in.window(1, 1, 10100, mybiomart, with.position=TRUE)
biomart.SNPid.in.window(1, 1, 10100, mybiomart, with.alleles=TRUE)
}
\author{
Alessia Visconti
}
