% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cut_p}
\alias{cut_p}
\title{Cut a continuous variable into given proportions}
\usage{
cut_p(x, p, ties.method = "random", fct_levels = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A numeric variable that is to be cut into categories}

\item{p}{The proportion of cases to be allocated to each category, in
ascending order. Should add up to one, otherwise, it will be scaled accordingly}

\item{ties.method}{Method for handling ties when ranking. Accepts "random"
(ties allocated randomly) or "in_order" (ties allocated in order of appearance).}

\item{fct_levels}{Character vector with names for levels. If it is NULL, the
groups will be labeled with their number and the cut-points employed.}

\item{verbose}{Should boundaries of groups be reported as message?}
}
\value{
Factor variable with x cut into length(p) categories in given
proportions
}
\description{
\code{cut()} and similar functions can cut continuous variables by quantile;
other helper functions exist to cut variables into groups of the same size
or width. This function cuts a continuous variable into given proportions.
}
\details{
When there are ties in the data that span group boundaries, this function
provides two methods for allocation: "random" (ties are allocated randomly)
and "in_order" (ties are allocated in the order they appear). The number of
observations per group is rounded up for even-numbered levels (second, fourth, etc)
and rounded down for others (except for the last level that is used to balance).
For large numbers of observations, the distribution will be very close to
what is desired, for very small numbers of observations, it should be checked.
}
\examples{
cut_p(iris$Sepal.Length, p = c(.25, .50, .25), fct_levels = c("short", "middling", "long"))
}
