% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dump_to_clip}
\alias{dump_to_clip}
\title{Dump objects to clipboard (to transfer them between R sessions)}
\usage{
dump_to_clip(objects)
}
\arguments{
\item{objects}{A character vector containing the names of one or more objects in the current session.}
}
\description{
This function calls dump on one or several R objects, which creates code that recreates them from the console.
It then copies this code to the clipboard. This can be used to quickly copy (small) objects between R sessions,
for instance during package development and testing, or - of course - to paste the dump code into a forum post.
}
