% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format confidence interval based on the bounds}
\usage{
fmt_ci(lower, upper, digits = 2)
}
\arguments{
\item{digits}{Number of significant digits, defaults to 2}

\item{x}{Numeric, or a vector of numbers}
}
\description{
Rounds numbers to consistent length and place in square brackets.
}
