% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_cor}
\alias{fmt_cor}
\title{Format number as correlation coefficient}
\usage{
fmt_cor(cor_value, digits = 2)
}
\arguments{
\item{digits}{Number of significant digits, defaults to 2}

\item{x}{Numeric, or a vector of numbers}
}
\description{
Format numbers as correlation coefficients, by rounding them and removing the leading 0
}
