% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format fraction as percentage string}
\usage{
fmt_pct(x, digits = 1)
}
\arguments{
\item{x}{Numeric, or a vector of numbers}

\item{digits}{Number of significant digits, defaults to 1}
}
\description{
Formats fractions as percentages, i.e. multiplying them by 100 and adding '%'
}
