% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_coef_rename_tribble}
\alias{get_coef_rename_tribble}
\title{Get code to generate tibbles to rename categorical variables and their levels}
\usage{
get_coef_rename_tribble(mod, show = TRUE)
}
\arguments{
\item{mod}{A model (e.g., returned from lm). coef() is called to extract the coefficient names.}

\item{show}{Logical - should the output be printed to the console. In any case, it is returned invisibly}
}
\value{
Code to be edited and passed to tibble::tribble() to create the coef_renames argument for
\code{\link{report_lm_with_std}} and similar functions. 'New' coefficient names are minimally
transformed, but are primarily intended to be placeholders.
}
\description{
Renaming categorical variables and their levels, for instance for summary tables, can be fiddly. This
function generates code in which only the new names need to be modified, and which can then be passed
to either \code{\link{rename_cat_variables}} or directly to \code{\link{cat_var_table}}
}
\details{
Only categorical variables should be passed to the function if code for levels is
requested. If a variable has more than 20 distinct values, it is dropped from the levels-tribble-code,
unless the max_levels argument is increased.
}
\examples{
mod <- lm(mpg ~ wt, mtcars)
get_coef_rename_tribble(mod)

}
