% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesaveR.R
\name{knit_print.timesaveR_raw_html}
\alias{knit_print.timesaveR_raw_html}
\title{Knitr S3 method to print tables}
\usage{
knit_print.timesaveR_raw_html(x, ...)
}
\arguments{
\item{x}{Object to knit_print}

\item{...}{Additional knit_print arguments}
}
\description{
These S3 methods are necessary to allow custom tables to print themselves in
knitr/rmarkdown documents.
}
