% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{mira.lm_F_test}
\alias{mira.lm_F_test}
\title{Report F-test for significance of multiply imputed lm models}
\usage{
mira.lm_F_test(mod, return_list = FALSE)
}
\arguments{
\item{mod}{A mira object (list of lm models in `analyses` element)}

\item{return_list}{Logical. Should items of test be returned in a list?
Otherwise, a string for reporting is returned, with Markdown formatting for APA style}
}
\description{
Takes a mira object (list of lm models based on mice imputations) and returns
an F-test for their significance, based on \code{\link[miceadds]{micombine.F}}
}
