% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{na_ifs}
\alias{na_ifs}
\title{Set variable to NA when it has specific values}
\usage{
na_ifs(x, replace)
}
\arguments{
\item{x}{The variable to transform.}

\item{replace}{One or more values to replace by NA.}
}
\description{
This function replaces specific values in a variable with NA. It would most commonly be used
to remove missing values indicated with codes such as "NA", -999 or "none".
(NB: It is very similar to dplyr's \code{na_if} but accepts more than one value.)
}
\examples{
library(dplyr)

ess_health \%>\% 
    mutate(eisced = na_ifs(eisced, c(7, 55)))
}
