% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{na_when}
\alias{na_when}
\title{Set variable to NA based on conditions}
\usage{
na_when(x, ..., operator = "|")
}
\arguments{
\item{x}{The variable to transform.}

\item{...}{One or more logical conditions, involving x or other variables.}

\item{operator}{Name of a logical operator to combine the conditions. Defaults to \code{"|"} (or),
\code{"&"} (and) is the other common choice, though \code{"xor"} would also work.}
}
\description{
This function sets a variable to NA based on one or several logical conditions.
It would most naturally be used inside a dplyr-mutate call.
By default, the conditions are combined with a logical OR, yet this can be changed
to AND by setting operator = "&".
}
\details{
Note that the function is called \code{na_when} to prevent clashing with dplyr's \code{na_if} ...
even though the latter might be the more intuitive name.
}
\keyword{"&"))}
\keyword{"DE",}
\keyword{%>%}
\keyword{19,}
\keyword{21))}
\keyword{6,}
\keyword{<}
\keyword{=}
\keyword{==}
\keyword{>}
\keyword{agea}
\keyword{cntry}
\keyword{cyl}
\keyword{ess_health}
\keyword{examples}
\keyword{library(dplyr)}
\keyword{mpg}
\keyword{mtcars}
\keyword{mutate(carb}
\keyword{mutate(eisced}
\keyword{na_when(carb,}
\keyword{na_when(eisced,}
\keyword{operator}
