% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{paired_t_test_d}
\alias{paired_t_test_d}
\title{Paired t.test with Cohen's d}
\usage{
paired_t_test_d(data, x, y)
}
\arguments{
\item{data}{A dataframe}

\item{x, y}{Character strings indicating the names of the two variables}
}
\value{
Invisibly returns a list including the t.test() output and Cohen's D
}
\description{
This function takes two variables that are representing paired data and
calculates a paired samples \code{t.test}. It then also calculates and prints
Cohen's d as a measure of effect size and shows a clearer data label than
the t.test function.
}
\examples{
paired_t_test_d(iris, "Sepal.Width", "Petal.Length")
}
