% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{print.timesaveR_raw_html}
\alias{print.timesaveR_raw_html}
\title{Renders HTML code for Viewer pane}
\usage{
\method{print}{timesaveR_raw_html}(x)
}
\arguments{
\item{x}{Either a character vector containing HTML code 
or a list with a html_code element}
}
\description{
Various functions in this package return HTML code, mostly for tables. This
function allows for them to rendered and shown in the Viewer. It can also be
called manually to render a character vector x that contains HTML code.
}
