% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_anova}
\alias{report_anova}
\title{Report (model comparison) ANOVA tests}
\usage{
report_anova(x)
}
\arguments{
\item{x}{An object returned from an anova() function}
}
\value{
A character string reporting the anova in APA style
}
\description{
This is under development and thus not yet exported.
}
\keyword{internal}
