% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{report_lm_with_std}
\alias{report_lm_with_std}
\title{Creates a summary table comparing standardised and non-standardised linear models}
\usage{
report_lm_with_std(
  mod,
  std_mod,
  conf_level = 0.95,
  fmt = "\%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  R2_change = FALSE,
  notes = list(NULL),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)
}
\arguments{
\item{mod}{A lm-model/mira object of lm models, with variables not standardised (or a list of such models)}

\item{std_mod}{A lm-model/mira object of lm models, with standardised variables. Can be
created with \code{\link{lm_std}} (or a list of such models)}

\item{conf_level}{Confidence level to use for confidence intervals, defaults to .95}

\item{fmt}{determines how to format numeric values
\itemize{
  \item integer: the number of digits to keep after the period `format(round(x, fmt), nsmall=fmt)`
  \item character: passed to the `sprintf` function (e.g., '\%.3f' keeps 3 digits with trailing zero). See `?sprintf`
  \item function: returns a formatted character string.
}}

\item{statistic_vertical}{Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{model_names}{If several pairs of models are to be plotted side by side, indicate the label for each *pair* here}

\item{show_nimp}{Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic}

\item{R2_change}{Logical. Report R2 change and F-test to compare models. Only implemented for comparing two pairs of models.}

\item{notes}{List of notes to append to bottom of table. An explanation of significance stars is automatically added. If the std models were run with a helper function in this package, a note regarding the standardisation is also automatically added.}

\item{apa_style}{Logical, should APA-style formatting be applied}

\item{stars}{Named vector of significance stars and their thresholds, check `rNuggets:::std_stars_pad` for default.}

\item{...}{
  Arguments passed on to \code{\link[modelsummary:modelsummary]{modelsummary::modelsummary}}
  \describe{
    \item{\code{output}}{filename or object type (character string)
\itemize{
  \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.
  \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
  \item Warning: the `output` argument \emph{cannot} be used when customizing tables with external packages. See the 'Details' section below.
}}
    \item{\code{estimate}}{string or `glue` string of the estimate to display (or a
vector with one string per model). Valid entries include any column name of
the data.frame produced by `get_estimates(model)`. Examples:
\itemize{
  \item "estimate"
  \item "{estimate} ({std.error}){stars}"
  \item "{estimate} [{conf.low}, {conf.high}]"
}}
    \item{\code{coef_map}}{named character vector. Values refer to the variable names
that will appear in the table. Names refer to the original term names stored
in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.
Coefficients that are omitted from this vector will be omitted from the
table. The table will be ordered in the same order as this vector.}
    \item{\code{coef_omit}}{string regular expression. Omits all matching coefficients
from the table using `grepl(perl=TRUE)`.}
    \item{\code{coef_rename}}{named character vector. Values refer to the variable names
that will appear in the table. Names refer to the original term names stored
in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.}
    \item{\code{gof_map}}{\itemize{
  \item NULL (default): the `modelsummary::gof_map` dictionary is used for formatting, and all unknown statistic are included.
  \item data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.
  \item list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.
}}
    \item{\code{gof_omit}}{string regular expression. Omits all matching gof statistics from
the table (using `grepl(perl=TRUE)`).}
    \item{\code{add_rows}}{a data.frame (or tibble) with the same number of columns as
your main table. By default, rows are appended to the bottom of the table.
You can define a "position" attribute of integers to set the row positions.
See Examples section below.}
    \item{\code{align}}{A character string of length equal to the number of columns in
the table.  "lcr" means that the first column will be left-aligned, the 2nd
column center-aligned, and the 3rd column right-aligned.}
    \item{\code{title}}{string}
  }}
}
\description{
This function creates a summary table for lm models (including mice::mira objects
containing lm-models) that shows a standardised and non-standardised version of the model
side-by-side. Several pairs of such models can be compared side-by-side.
}
\examples{
ess_health$gndr <- factor(ess_health$gndr)

#Standard lm model
mod1 <- lm(depression ~ agea + gndr + health + cntry, ess_health)

#Model with standardised coefficients

mod2 <- lm_std(depression ~ agea + gndr + health + cntry, ess_health)

report_lm_with_std(mod1, mod2)
}
