% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{round_df}
\alias{round_df}
\title{Round all numeric columns in dataframe}
\source{
https://stackoverflow.com/questions/9063889/how-to-round-a-data-frame-in-r-that-contains-some-character-variables
}
\usage{
round_df(df, digits = 2)
}
\arguments{
\item{df}{Dataframe to be rounded}

\item{digits}{Number of digits, defaults to 2}
}
\description{
Rounds all numeric columns in dataframe, using the R default "half to even"
}
