% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{run_and_format}
\alias{run_and_format}
\title{Run a code snippet and copy formatted code and output to clipboard}
\usage{
run_and_format(code = NULL)
}
\arguments{
\item{code}{The code to run - can be NULL, then the code is retrieved from the clipboard.}
}
\description{
This builds on the \code{\link[reprex:reprex]{reprex::reprex()}} function to take a chunk of code,
run it and copy the formatted code and output to the clipboard. Note that
this does NOT run in a clean session and is thus unlikely to create \emph{reproducible}
examples. Instead, it is intended to be used to quickly get a formatted bit
of code and output that is not self-contained, e.g., for teaching materials.
}
\examples{
if (interactive()) {
  name <- "Lukas"
  run_and_format(paste("My name is", name))
}
}
