% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scales.R
\name{svy_make_scale}
\alias{svy_make_scale}
\title{Create scale by calculating item mean and returns descriptives for srvyr objects}
\usage{
svy_make_scale(
  data,
  scale_items,
  scale_name,
  print_hist = TRUE,
  print_desc = TRUE,
  scale_title = scale_name,
  reversed = NULL,
  r_key = NULL
)
}
\arguments{
\item{data}{A srvyr survey object}

\item{scale_items}{A characters vector containing the items for that scale
(variables in data)}

\item{scale_name}{Character. The name of the variable the scale should be saved as}

\item{print_hist}{Logical. Should histograms of the scale and its items be printed.}

\item{print_desc}{Logical. Should descriptive statistics for the scale be printed.}

\item{scale_title}{Character. Name of scale for printing. Defaults to scale_name}

\item{reversed}{(optional) A characters vector containing the items that should be reverse-coded (
subset of scale_items)}

\item{r_key}{(optional) Numeric. Set to the possible maximum value of the scale
if the whole scale should be reversed, or to -1 to reverse the scale based on
the observed maximum}
}
\value{
The survey object with the scale added as an additional variable.
}
\description{
This function creates a scale by calculating the mean of a set of items,
and prints and returns descriptives that allow to assess internal consistency
and spread. It is primarily based on the \code{psych::alpha} function, with
more parsimonious output and some added functionality.
}
\examples{

library(survey)
data(api)

svy_data <- svydesign(id = ~1, strata = ~stype, weights = ~pw, 
                   data = apistrat, fpc = ~fpc)
scale_items <- c("ell", "meals", "mobility", "col.grad", "full")
scale_reversed <- c("col.grad", "full")                    
svy_make_scale(svy_data, scale_items, "SES", reversed = scale_reversed)    
}
