% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_helpers.R
\name{svy_miss_var_summary}
\alias{svy_miss_var_summary}
\title{Create overview over missing data in survey object}
\usage{
svy_miss_var_summary(
  svy_data,
  ...,
  .any_missing = TRUE,
  .include_complete = FALSE
)
}
\arguments{
\item{svy_data}{A survey object created with the survey package}

\item{...}{Variables to consider. By default this looks at the whole dataset. 
Otherwise, this should be one or more unquoted expressions separated by commas, or
a tidyselect helper function (e.g., `starts_with()` or `where(is.numeric)`).}

\item{.any_missing}{Should a row be shown at the start of the return with the
number and share of (weighted) responses that have missing data on at least one
of the variables considered? Defaults to TRUE.}

\item{.include_complete}{Should variables with no missing data be included
in the returned tibble? Defaults to FALSE.}
}
\description{
This function creates a tibble that summarises missing data 
with survey weights for missing counts and proportions. (Apart from
that, it recreates the \code{naniar::miss_var_summary()}
output.)
}
\examples{
library(survey)
data(api)
svy_df <- svydesign(id = ~1, strata = ~stype, weights = ~pw, 
                   data = apistrat, fpc = ~fpc)

svy_miss_var_summary(svy_df, acs.core, target, name)    
}
