% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesaveR.R
\docType{package}
\name{timesaveR}
\alias{timesaveR}
\alias{timesaveR-package}
\title{Functions to Accelerate (Academic) Data Analysis and Reporting}
\description{
Functions and templates in this package facilitate common tasks
in the course of most research projects in social psychology and related
fields. These include creating scales, reporting descriptive statistics,
correlations and distributions, reporting linear regression models with
standardized coefficients and F-change comparisons, as well as plotting
mediation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lukaswallrich.github.io/timesaveR/}
  \item \url{https://lukaswallrich.github.io/timesaveR}
  \item Report bugs at \url{https://github.com/LukasWallrich/timesaveR/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com}

}
